requirements.txt
alembic==1.7.6
aniso8601==9.0.1
astroid==2.9.3
bcrypt==3.2.0
cffi==1.15.0
click==8.0.3
colorama==0.4.4
Flask==2.0.2
flask-marshmallow==0.14.0
Flask-Migrate==3.1.0
Flask-RESTful==0.3.9
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
marshmallow-sqlalchemy==0.27.0
mccabe==0.6.1
passlib==1.7.4
platformdirs==2.4.1
psycopg2==2.9.3
pycodestyle==2.8.0
pycparser==2.21
pylint==2.12.2
python-dotenv==0.19.2
pytz==2021.3
six==1.16.0
SQLAlchemy==1.4.31
toml==0.10.2
Werkzeug==2.0.2
wrapt==1.13.3
